[Refactor] 백업파일 삭제, 수정된 파일 index.tsx 로 이름 변경
